Break ShareSave into several smaller components for managability.
